% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drill_spec.r
\name{drill_spec}
\alias{drill_spec}
\title{Drill a timeseries of a variable at a given location}
\usage{
drill_spec(input, spec, lat, lon)
}
\arguments{
\item{input}{An open ncdf4 object.}

\item{spec}{Full species name (4 chars) including spaces.}

\item{lat}{Latitude of sample point.}

\item{lon}{Longitude of sample point}
}
\value{
Timeseries vector of species concentration.
}
\description{
Drill a timeseries of a variable at a given location
}
\examples{
add(1, 1)
tfile <- system.file("extdata","aust_20140212_nat_0.800dg_morw.nc",package="raaqfs")
mydata <- nc_open(tfile)
lat <- -33.5
lon <- 151.0
timeseries <- drill_spec(mydata,"EC25",lat,lon)
plot(timeseries,type="l",xlab="Hour",ylab="conc")
}

