% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spec_list.r
\name{get_spec_list}
\alias{get_spec_list}
\title{Get species list from NetCDF file}
\usage{
get_spec_list(input)
}
\arguments{
\item{input}{An open ncdf4 object.}
}
\value{
The vector of chemical species codes.
}
\description{
Get species list from NetCDF file
}
\examples{
add(1, 1)
tfile <- system.file("extdata","aust_20140212_nat_0.800dg_morw.nc",package="raaqfs")
mydata <- nc_open(tfile)
species<-get_spec_list(mydata)
print(species)
}

