% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_spec.r
\name{stack_spec}
\alias{stack_spec}
\title{Return raster stack of variable for a day.}
\usage{
stack_spec(input, spec)
}
\arguments{
\item{input}{An open ncdf4 object.}

\item{spec}{Full species name (4 chars) including spaces.}
}
\value{
RasterStack of concentration.
}
\description{
Return raster stack of variable for a day.
}
\examples{
library(ncdf4)
library(raster)
tfile <- system.file("extdata","aust_20140212_nat_0.800dg_morw.nc",package="raaqfs")
mydata <- nc_open(tfile)
salt <- stack_spec(mydata,"SS25") 
animate(salt,pause=0.1,n=2)
}

